+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed
 - Bumped minimum hikari version to 2.0.0.dev108.
+- Use `None` as the default for `max_value` and `min_value` in `parsing.py`.
 
 ## [2.4.3a1] - 2022-03-12
 ### Added